add scss and reorganize folders
