+++ client/Routes.js
@@ -1,7 +1,7 @@
 import React, {Component, Fragment} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch, Redirect} from 'react-router-dom'
-import { Login, Signup } from './components/AuthForm';
+import { Login, Signup } from './components/AuthForm/AuthForm';
 import Home from './components/Home';
 import CandyList from './components/CandyList';
 import Candy from './components/Candy';
