+++ client/components/AuthForm/AuthForm.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {connect} from 'react-redux'
-import {authenticate} from '../store'
+import {authenticate} from '../../store'
 
 /**
  * COMPONENT
