+++ client/components/Candy/index.js
@@ -1,0 +1,3 @@
+import Candy from './Candy';
+
+export default Candy;
