+++ client/components/CandyList/CandyList.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
-import { getCandy } from "../store/candy";
+import { getCandy } from "../../store/candy";
 
 class CandyList extends React.Component {
