+++ client/components/CandyList/index.js
@@ -1,0 +1,3 @@
+import CandyList from './CandyList';
+
+export default CandyList;
