+++ client/components/Cart/index.js
@@ -1,0 +1,3 @@
+import Cart from './Cart';
+
+export default Cart;
